add phong in ray tracing equation

+++ src/texture.ml
@@ -2,7 +2,7 @@ type t = {
 	color : Color.t; 
 	kd : float;
 	ks : float;
-	n : float;
+	n : int;
 }
 
 let make color kd ks n = {color; kd; ks; n}
